🪚 Extract chat logic and views into separate components
Split ChatView into ChatMessagesView, ChatInputView and ChatLogicHandler to improve maintainability
